+++ lib/playwright_ex/processes/frame_state.ex
@@ -1,0 +1,37 @@
+defmodule PlaywrightEx.FrameState do
+  @moduledoc false
+
+  defstruct url: "", load_states: MapSet.new()
+
+  @type t :: %__MODULE__{url: String.t(), load_states: MapSet.t(String.t())}
+
+  def new(initializer) do
+    %__MODULE__{
+      url: initializer[:url] || "",
+      load_states: MapSet.new(List.wrap(initializer[:load_states]), &to_string/1)
+    }
+  end
+
+  def update(state, :loadstate, params) do
+    load_states =
+      state.load_states
+      |> add(params[:add])
+      |> remove(params[:remove])
+
+    %{state | load_states: load_states}
+  end
+
+  def update(state, :navigated, params) do
+    load_states = if Map.has_key?(params, :new_document), do: MapSet.new(["commit"]), else: state.load_states
+    %{state | url: params[:url] || state.url, load_states: load_states}
+  end
+
+  def error(:frame_detached), do: {:error, %{message: "Navigating frame was detached!"}}
+  def error(:page_closed), do: {:error, %{message: "Navigation failed because page was closed!"}}
+  def error(:page_crashed), do: {:error, %{message: "Navigation failed because page crashed!"}}
+
+  defp add(states, nil), do: states
+  defp add(states, value), do: MapSet.put(states, to_string(value))
+  defp remove(states, nil), do: states
+  defp remove(states, value), do: MapSet.delete(states, to_string(value))
+end